This commit changes the footer elements so that they have links around them in the mobile version of the homepage for the project. These elements were previously created using divs, which were stacked using the CSS flex property. A mouse effect was used to create the appearance of a link around these elements. This commit modifies the footer elements, making them list items instead of the previously used flexboxed divs. This change aims to establish a more systematic approach when using <a> tags around them.
The second change is the addition of images to the webpage. Previous commits modified the file structure of the project, allowing these images to be added in HTML, which this commit implements. The heart icons have also been replaced with icons from the Font Awesome library, rather than exported SVGs from Figma, to enable their interactivity.